services/local_ai/requirements.txt
# Local AI Service Dependencies - Optimized for Performance & Quality

# Core Framework
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-multipart==0.0.6
requests==2.31.0

# Configuration Management
PyYAML==6.0.1  # YAML-driven configuration
python-dotenv==1.0.1  # Environment variable management

# AI/ML Stack - Latest Optimized Versions
torch==2.2.2  # Removed +cpu suffix for compatibility
transformers==4.40.1  # Latest with performance improvements
accelerate==0.29.3  # Better memory management
bitsandbytes==0.43.1  # Improved quantization
gpt4all==2.7.0  # Latest with model optimizations

# Performance Optimization
optimum==1.19.1  # Intel/ONNX optimizations
onnxruntime==1.17.3  # Faster inference
sentence-transformers==2.7.0  # Optimized embeddings

# Model Management & Caching
huggingface-hub==0.23.0  # Better download management
cachier==3.0.1  # Smart caching for model downloads
diskcache==5.6.3  # Persistent model caching

# Memory & Performance Monitoring
psutil==5.9.8  # System resource monitoring
numpy==1.26.4  # Optimized linear algebra
einops==0.8.0  # Efficient tensor operations